chore: refactor http response binding protocol
