+++ codegen/smithy-aws-swift-codegen/src/main/kotlin/software/amazon/smithy/aws/swift/codegen/awsjson/AwsJson1_0_ProtocolGenerator.kt
@@ -17,10 +17,10 @@ open class AwsJson1_0_ProtocolGenerator : AWSHttpBindingProtocolGenerator() {
     override val httpProtocolClientGeneratorFactory = AWSHttpProtocolClientGeneratorFactory()
     override val httpProtocolCustomizable = AWSHttpProtocolClientCustomizableFactory().constructClientCustomizable(protocol)
     override val httpResponseGenerator = HttpResponseGenerator(
-        AWSHttpResponseBindingAWSJson(),
         serviceErrorProtocolSymbol,
         unknownServiceErrorSymbol,
-        defaultTimestampFormat
+        defaultTimestampFormat,
+        AWSJsonHttpResponseBindingErrorGenerator()
     )
 
     override fun getProtocolHttpBindingResolver(ctx: ProtocolGenerator.GenerationContext):
