+++ codegen/smithy-aws-swift-codegen/src/main/kotlin/software/amazon/smithy/aws/swift/codegen/restxml/AWSRestXMLHttpResponseBindingErrorGenerator.kt
@@ -1,4 +1,4 @@
-package software.amazon.smithy.aws.swift.codegen.restjson
+package software.amazon.smithy.aws.swift.codegen.restxml
 
 import software.amazon.smithy.aws.swift.codegen.AWSSwiftDependency
 import software.amazon.smithy.codegen.core.Symbol
@@ -8,8 +8,8 @@ import software.amazon.smithy.swift.codegen.defaultName
 import software.amazon.smithy.swift.codegen.integration.ProtocolGenerator
 import software.amazon.smithy.swift.codegen.integration.httpResponse.HttpResponseBindingErrorGeneratable
 
-class AWSHttpResponseBindingRestJson : HttpResponseBindingErrorGeneratable {
-    override fun renderHttpResponseBinding(ctx: ProtocolGenerator.GenerationContext, op: OperationShape) {
+class AWSRestXMLHttpResponseBindingErrorGenerator : HttpResponseBindingErrorGeneratable {
+    override fun render(ctx: ProtocolGenerator.GenerationContext, op: OperationShape) {
         val operationErrorName = "${op.defaultName()}OutputError"
         val rootNamespace = ctx.settings.moduleName
         val httpBindingSymbol = Symbol.builder()
@@ -18,14 +18,13 @@ class AWSHttpResponseBindingRestJson : HttpResponseBindingErrorGeneratable {
             .build()
 
         ctx.delegator.useShapeWriter(httpBindingSymbol) { writer ->
-            writer.addImport(AWSSwiftDependency.AWS_CLIENT_RUNTIME.target)
-            writer.addImport(SwiftDependency.CLIENT_RUNTIME.target)
+            writer.addImport(AWSSwiftDependency.AWS_CLIENT_RUNTIME.packageName)
+            writer.addImport(SwiftDependency.CLIENT_RUNTIME.packageName)
 
             writer.openBlock("extension \$L: HttpResponseBinding {", "}", operationErrorName) {
                 writer.openBlock("public init(httpResponse: HttpResponse, decoder: ResponseDecoder? = nil) throws {", "}") {
-                    writer.write("let errorDetails = try RestJSONError(httpResponse: httpResponse)")
-                    writer.write("let requestID = httpResponse.headers.value(for: X_AMZN_REQUEST_ID_HEADER)")
-                    writer.write("try self.init(errorType: errorDetails.errorType, httpResponse: httpResponse, decoder: decoder, message: errorDetails.errorMessage, requestID: requestID)")
+                    writer.write("let errorDetails = try RestXMLError(httpResponse: httpResponse)")
+                    writer.write("try self.init(errorType: errorDetails.errorCode, httpResponse: httpResponse, decoder: decoder, message: nil, requestID: errorDetails.requestId)")
                 }
             }
         }
