+++ codegen/smithy-aws-swift-codegen/src/test/kotlin/software/amazon/smithy/aws/swift/codegen/mocks/MockRestJson1BindingProtocolGenerator.kt
@@ -3,7 +3,7 @@ package software.amazon.smithy.aws.swift.codegen.mocks
 import software.amazon.smithy.aws.swift.codegen.AWSHttpBindingProtocolGenerator
 import software.amazon.smithy.aws.swift.codegen.AWSHttpProtocolClientCustomizableFactory
 import software.amazon.smithy.aws.swift.codegen.AWSHttpProtocolClientGeneratorFactory
-import software.amazon.smithy.aws.swift.codegen.restjson.AWSHttpResponseBindingRestJson
+import software.amazon.smithy.aws.swift.codegen.restjson.AWSRestJson1HttpResponseBindingErrorGeneratable
 import software.amazon.smithy.aws.traits.protocols.RestJson1Trait
 import software.amazon.smithy.model.shapes.ShapeId
 import software.amazon.smithy.model.traits.TimestampFormatTrait
@@ -16,9 +16,9 @@ class MockRestJson1BindingProtocolGenerator : AWSHttpBindingProtocolGenerator() 
     override val httpProtocolClientGeneratorFactory = AWSHttpProtocolClientGeneratorFactory()
     override val httpProtocolCustomizable = AWSHttpProtocolClientCustomizableFactory().constructClientCustomizable(protocol)
     override val httpResponseGenerator = HttpResponseGenerator(
-        AWSHttpResponseBindingRestJson(),
         serviceErrorProtocolSymbol,
         unknownServiceErrorSymbol,
-        defaultTimestampFormat
+        defaultTimestampFormat,
+        AWSRestJson1HttpResponseBindingErrorGeneratable()
     )
 }
